Change rule to AssertEqualsArgumentOrder
