+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/internal/TestFrameworksUtil.java
@@ -7,6 +7,8 @@ package net.sourceforge.pmd.lang.java.rule.internal;
 import static net.sourceforge.pmd.util.CollectionUtil.setOf;
 import static net.sourceforge.pmd.util.CollectionUtil.union;
 
+import java.util.Arrays;
+import java.util.List;
 import java.util.Set;
 
 import net.sourceforge.pmd.lang.java.ast.ASTAnnotation;
@@ -14,11 +16,13 @@ import net.sourceforge.pmd.lang.java.ast.ASTClassDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTMethodCall;
 import net.sourceforge.pmd.lang.java.ast.ASTMethodDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTTypeDeclaration;
+import net.sourceforge.pmd.lang.java.ast.InvocationNode;
 import net.sourceforge.pmd.lang.java.ast.JModifier;
 import net.sourceforge.pmd.lang.java.ast.ModifierOwner.Visibility;
 import net.sourceforge.pmd.lang.java.symbols.JClassSymbol;
 import net.sourceforge.pmd.lang.java.symbols.JMethodSymbol;
 import net.sourceforge.pmd.lang.java.symbols.JTypeDeclSymbol;
+import net.sourceforge.pmd.lang.java.types.InvocationMatcher;
 import net.sourceforge.pmd.lang.java.types.JClassType;
 import net.sourceforge.pmd.lang.java.types.JTypeMirror;
 import net.sourceforge.pmd.lang.java.types.TypeTestUtil;
@@ -27,6 +31,23 @@ import net.sourceforge.pmd.lang.java.types.TypeTestUtil;
  * Utilities for rules related to test frameworks (Junit, TestNG, etc).
  */
 public final class TestFrameworksUtil {
+
+
+    public static final class EqualMethod {
+        public final int actualPosition;
+        public final int expectedPosition;
+        private final InvocationMatcher matcher;
+
+        private EqualMethod(String pattern, int actualPosition, int expectedPosition) {
+            this.matcher = InvocationMatcher.parse(pattern);
+            this.expectedPosition = expectedPosition;
+            this.actualPosition = actualPosition;
+        }
+
+        public boolean matches(InvocationNode node) {
+            return matcher.matchesCall(node);
+        }
+    }
 
     private static final String JUNIT3_CLASS_NAME = "junit.framework.TestCase";
     private static final String JUNIT4_TEST_ANNOT = "org.junit.Test";
@@ -86,6 +107,23 @@ public final class TestFrameworksUtil {
                     JUNIT4_CONFIGURATION_ANNOTATIONS,
                     TEST_NG_CONFIGURATION_ANNOTATIONS
             );
+
+    public static final List<EqualMethod> EQUAL_METHODS = Arrays.asList(
+        // JUnit Jupiter: expected, actual, [message]
+        new EqualMethod("org.junit.jupiter.api.Assertions#assertEquals(_,_)", 1, 0),
+        new EqualMethod("org.junit.jupiter.api.Assertions#assertEquals(_,_,_)", 1, 0),
+        // JUnit 3 and 4, Spring: [message], expected, actual
+        new EqualMethod("org.junit.Assert#assertEquals(_,_)", 1, 0),
+        new EqualMethod("org.junit.Assert#assertEquals(_,_,_)", 2, 1),
+        new EqualMethod("junit.framework.TestCase#assertEquals(_,_)", 1, 0),
+        new EqualMethod("junit.framework.TestCase#assertEquals(_,_,_)", 2, 1),
+        new EqualMethod("org.springframework.test.util.AssertionErrors#assertEquals(_,_,_)", 2, 1),
+        // TestNG: actual, expected, [message]
+        new EqualMethod("org.testng.Assert#assertEquals(_*)", 0, 1),
+        // JSONAssert: [message], expected, actual, compare mode
+        new EqualMethod("org.skyscreamer.jsonassert.JSONAssert#assertEquals(_,_,_)", 1, 0),
+        new EqualMethod("org.skyscreamer.jsonassert.JSONAssert#assertEquals(_,_,_,_)", 2, 1)
+    );
 
     private TestFrameworksUtil() {
         // utility class
