+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/rule/errorprone/xml/AssertEqualsArgumentOrder.xml
@@ -6,8 +6,8 @@
 
     <test-code>
         <description>JUnit Jupiter</description>
-        <expected-problems>3</expected-problems>
-        <expected-linenumbers>6,7,8</expected-linenumbers>
+        <expected-problems>1</expected-problems>
+        <expected-linenumbers>8</expected-linenumbers>
         <code>
             <![CDATA[
 import org.junit.jupiter.api.Test;
@@ -15,8 +15,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 public class SimpleTest {
     @Test
     public void testX() {
-        assertEquals("foo", "bar", "reason"); // this could be just fail("reason")
-        assertEquals("foo", "foo", "other reason"); // this should be removed
+        assertEquals("foo", "bar", "reason"); // not covered by this rule
+        assertEquals("foo", "foo", "other reason"); // not covered by this rule
         assertEquals(next("foo"), "bar"); // wrong argument order
         assertEquals("bar", next("foo")); // good
     }
@@ -30,8 +30,8 @@ public class SimpleTest {
     </test-code>
     <test-code>
         <description>JUnit 4</description>
-        <expected-problems>3</expected-problems>
-        <expected-linenumbers>6,7,8</expected-linenumbers>
+        <expected-problems>1</expected-problems>
+        <expected-linenumbers>8</expected-linenumbers>
         <code>
             <![CDATA[
 import org.junit.Test;
@@ -39,8 +39,8 @@ import static org.junit.Assert.assertEquals;
 public class SimpleTest {
     @Test
     public void testX() {
-        assertEquals("foo", "bar", "reason"); // this could be just fail("reason")
-        assertEquals("foo", "foo", "other reason"); // this should be removed
+        assertEquals("foo", "bar", "reason"); // not covered by this rule
+        assertEquals("foo", "foo", "other reason"); // not covered by this rule
         assertEquals(next("foo"), "bar"); // wrong argument order
         assertEquals("bar", next("foo")); // good
     }
@@ -55,8 +55,8 @@ public class SimpleTest {
 
     <test-code>
         <description>TestNG</description>
-        <expected-problems>3</expected-problems>
-        <expected-linenumbers>6,7,8</expected-linenumbers>
+        <expected-problems>1</expected-problems>
+        <expected-linenumbers>8</expected-linenumbers>
         <code>
             <![CDATA[
 import org.testng.annotations.Test;
@@ -64,8 +64,8 @@ import org.testng.Assert;
 public class SimpleTest {
     @Test
     public void testX() {
-        Assert.assertEquals("foo", "bar", "reason"); // this could be just fail("reason")
-        Assert.assertEquals("foo", "foo", "other reason"); // this should be removed
+        Assert.assertEquals("foo", "bar", "reason"); // not covered by this rule
+        Assert.assertEquals("foo", "foo", "other reason"); // not covered by this rule
         Assert.assertEquals("bar", next("foo")); // wrong argument order
         Assert.assertEquals(next("foo"), "bar"); // good
     }
@@ -86,7 +86,7 @@ public class SimpleTest {
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 public class SimpleTest {
-    private static String CONST = "foo";
+    private static final String CONST = "foo";
     @Test
     public void testX() {
         assertEquals(CONST, "bar", "reason"); // wrong argument order
